edit

+++ routers/admin/cardType.js
@@ -6,7 +6,16 @@ const router = express.Router();
 // Create CardType
 router.post("/", adminAuth, async (req, res) => {
   const { image, name, companyCardID } = req.body;
+  const userType = req.user.type;
+  const permisson = req.user.permissons || {};
+
+ 
   try {
+
+    if (userType == 'ADMIN' && !permisson.create_card_type) {
+    return res.status(400).json({ error: "No permission to create card type" });
+
+  }
     const cardType = await prisma.cardType.create({
       data: { image, name, companyCardID },
     });
@@ -18,7 +27,14 @@ router.post("/", adminAuth, async (req, res) => {
 
 // Read all CardTypes
 router.get("/", adminAuth, async (req, res) => {
+  const userType = req.user.type;
+  const permisson = req.user.permissons || {};
+
   try {
+
+    if (userType == 'ADMIN' && !permisson.read_card_type) {
+      return res.status(400).json({ error: "No permission to read card type" });
+    }
     const take = parseInt(req.query.take || 8);
     const skip = parseInt(req.query.skip || 0);
     const q = req.query.q || "";
@@ -52,7 +68,16 @@ router.get("/", adminAuth, async (req, res) => {
 // Read CardType by ID
 router.get("/:id", async (req, res) => {
   const { id } = req.params;
+  const userType = req.user.type;
+  const permisson = req.user.permissons || {};
+
+  
   try {
+
+    if (userType == 'ADMIN' && !permisson.read_card_type) {
+      return res.status(400).json({ error: "No permission to read card type" });
+    }
+
     const cardType = await prisma.cardType.findUnique({
       where: { id: Number(id) },
     });
@@ -66,7 +91,14 @@ router.get("/:id", async (req, res) => {
 router.put("/:id", adminAuth, async (req, res) => {
   const { id } = req.params;
   const { image, name, companyCardID } = req.body;
+  const userType = req.user.type;
+  const permisson = req.user.permissons || {};
+
   try {
+
+    if (userType == 'ADMIN' && !permisson.update_card_type) {
+      return res.status(400).json({ error: "No permission to update card type" });
+    }
     const cardType = await prisma.cardType.update({
       where: { id: Number(id) },
       data: { image, name, companyCardID },
@@ -80,7 +112,14 @@ router.put("/:id", adminAuth, async (req, res) => {
 router.put("/active/:id", adminAuth, async (req, res) => {
   const { id } = req.params;
   const { active } = req.body;
+  const userType = req.user.type;
+  const permisson = req.user.permissons || {};
+
   try {
+
+    if (userType == 'ADMIN' && !permisson.card_type_status) {
+      return res.status(400).json({ error: "No permission to update card type status" });
+    }
     const cardType = await prisma.cardType.update({
       where: { id: Number(id) },
       data: { active },
@@ -94,7 +133,14 @@ router.put("/active/:id", adminAuth, async (req, res) => {
 // Delete CardType by ID
 router.delete("/:id", adminAuth, async (req, res) => {
   const { id } = req.params;
+  const userType = req.user.type;
+  const permisson = req.user.permissons || {};
+
   try {
+
+    if (userType == 'ADMIN' && !permisson.delete_card_type) {
+      return res.status(400).json({ error: "No permission to delete card type" });
+    }
     const cardType = await prisma.cardType.delete({
       where: { id: Number(id) },
     });
